requirements.txt
torch
torchvision
Pillow
open_clip_torch
matplotlib
numpy
tqdm
